Fix custom roll formulas on Moves
A section of the roll handling code for moves attempts to update the
output template with the stat and stat value of the roll, but this
information doesn't exist when handling a custom roll formula. This
would lead to an 'undefined' exception when trying to access the stat. The
changes in this commit check to see if the stat data exist before
accessing it to update the template text.