Add bounds options
Some PbtA games impose bounds on total amount you can modify a roll.
This adds support for including that in a game's TOML as:

 - minMod
 - maxMod

If either of those are missing, this logic will be skipped, and you can
modify rolls to the moon and back.

The primary example of a game that clamps rolls that I'm thinking of is
Masks, which (on p32) says that, all modifiers inclusive, you can't roll
at more than +4 or less than -3.